python_streamlit_demo/requirements.txt
streamlit>=1.38,<2
pandas>=2.0.0
databricks-sql-connector>=3.0.0
